PERCEPTION: Rename package to navigator_vison
also fix somethings outputed by DDB -Wall

+++ perception/navigator_vision/nodes/shape_identification/main.cpp
@@ -1,23 +1,18 @@
 #include "std_msgs/String.h"
-
 #include <cv_bridge/cv_bridge.h>
 #include <image_transport/image_transport.h>
 #include <ros/ros.h>
 #include <sensor_msgs/image_encodings.h>
-
-#include "opencv2/core/core.hpp"
-#include "opencv2/highgui/highgui.hpp"
-#include "opencv2/imgproc/imgproc.hpp"
-#include "opencv2/opencv.hpp"
-
+#include <opencv2/core/core.hpp>
+#include <opencv2/highgui/highgui.hpp>
+#include <opencv2/imgproc/imgproc.hpp>
+#include <opencv2/opencv.hpp>
 #include <navigator_msgs/DockShapes.h>
-
 #include "GrayscaleContour/GrayscaleContour.h"
-#include "std_srvs/SetBool.h"
-//#include "FeatureDetectorMethod/FeatureDetectorMethod.h"
+#include <std_srvs/SetBool.h>
 #include <navigator_msgs/SetROI.h>
 #include "DockShapeVision.h"
-#include "sensor_msgs/RegionOfInterest.h"
+#include <sensor_msgs/RegionOfInterest.h>
 #include "ShapeTracker.h"
 
 
@@ -27,7 +22,6 @@ class ShooterVision {
  private:
   ShapeTracker tracker;
   std::unique_ptr<DockShapeVision> vision;
-  // ros frame thing
   navigator_msgs::DockShapes symbols;
   ros::NodeHandle nh_;
   image_transport::ImageTransport it_;
@@ -106,7 +100,7 @@ class ShooterVision {
     symbols.list.clear();
 
     vision->GetShapes(frame, roi, symbols);
-    for (unsigned int i = 0; i < symbols.list.size(); i++)
+    for (size_t i = 0; i < symbols.list.size(); i++)
       symbols.list[i].header = msg->header;
     foundShapesPublisher.publish(symbols);
     tracker.addShapes(symbols);
